Read Description
ConversationChatGPTwithGUI.py: This script creates an interactive web interface using Gradio, where users can converse with the GPT-3.5-turbo model by OpenAI in the role of a real estate investment expert. The assistant enhances and builds upon the user's input and provides a corresponding response.

ImageGenerator.py: This script generates images based on a description provided by the user. It interacts with the OpenAI API, specifically a model like DALL-E, to generate and save images as per the user's descriptions. The number of iterations can be determined by the user.

NonConversationChatGPT.py: This script accepts a user's textual description, enhances it using OpenAI's DaVinci engine and outputs the enhanced description. The role of the AI here is to convert the user's description into a more detailed one, suitable for generating an image.